# service/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:5000
info:
  contact: {}
  description: Backend API for MimiCare Project (Gin + MongoDB)
  title: MimiCare API
  version: "1.0"
paths:
...
  /notification/get:
    get:
      description: ดึงการแจ้งเตือนทั้งหมดของผู้ใช้ที่ล็อกอินอยู่ (ผู้ใช้ต้องเป็นเจ้าของการแจ้งเตือนเหล่านี้)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ดึงการแจ้งเตือนทั้งหมดของผู้ใช้ตาม ID
      tags:
      - Notification
  /notification/read-all:
    put:
      description: ทำเครื่องหมายการแจ้งเตือนทั้งหมดว่าอ่านแล้ว (สำหรับผู้ใช้ที่ล็อกอินอยู่)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      summary: ทำเครื่องหมายการแจ้งเตือนทั้งหมดว่าอ่านแล้ว
      tags:
      - Notification
  /notification/read/{id}:
    put:
      description: ทำเครื่องหมายการแจ้งเตือนว่าอ่านแล้วตาม ID ของการแจ้งเตือน (ผู้ใช้ต้องเป็นเจ้าของการแจ้งเตือนนี้)
      parameters:
      - description: ObjectID ของการแจ้งเตือน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ทำเครื่องหมายการแจ้งเตือนว่าอ่านแล้วตาม ID
      tags:
      - Notification
...
schemes:
- http
swagger: "2.0"